Implemented #335
